[K8S] Factored out Matsuri::Concerns::PodTemplate
